refactor: move chart navigation controls into the top toolbar
This commit moves the chart navigation controls (selection, zoom-in, zoom-out, zoom-pan, and scale to fit) into the top chart toolbar, effectively removing the display bar from the right-hand side of the chart. As such, the former "ChartFilterControlBar" has been renamed to "ChartControlBar", and the former "ChartDisplayControlBar" has been changed to "ChartButtonGroup" and is now a composite that only holds the buttons.