feat: move lane resize controls to control bar and restore old zoom listener for overview
This commit moves the lane resize controls to the main control bar for the chart. Additionally, when the chart is zoomed out to show an overview of the lanes, the old zoom listener is attached to the mouse wheel. When overview mode is toggled off the last used lane height is restored.