feat: replace popup table with foldable sashform
While an interesting feature, trying to persist settings and selections from the FilterEditor and table within the former popup table was not a realistic feat. After demoing the threads page work at the latest bi-weekly JMC hangout meeting, it was noted that it'd be nice to be able to easily interact with the table and update selections and filters. Based on this information, this commit replaces the popup table with a foldable sashform using controls proposed in JMC-6386. This allows the user to quickly hide and unhide both the table or chart based on their workflow.