6936: Parser should track statistics during parsing
Add new IParserStats interface implemented by EventCollection.
That way, IItemCollection is not touched, but you need to cast the
instance to IParserStats.
Support of old (v0.x) JFR format.
During parsing, we collect number of events per type, the size, number
of chunks and version of the file.
We also track for version 1.x/2.x the number of skipped events.
RecordPrinter tool class is modified to support a summary verbosity mode
and -summary option to print statistics per event type.