Remove dependency on jetty.servlet-api, use jakarta instead
We needed a servlet api implementation for the websocket server
so we added a dependency on the jetty.servlet-api. We already
have jakarta.servlet-api in our dependencies, required by the
eclipse help page which also uses jetty to serve dynamic content.

We need to avoid loading multiple servlet-apis because it causes
errors serving help pages - the server fails to recognise HTTP
requests, presumably because of the multiple implementations available.