7806: Add an export option for Stacktraces in collapsed format
Add a new context menu when clicking on stacktrace view to export
selected stacktraces as collapsed format. With this format it's
easy to manipulate/filter stacktraces as text file and generate
a flamegraph with the perl script from the flamegraph repository
(https://github.com/brendangregg/FlameGraph)

The generated content is copied into the clipboard.