Implement scaffolding for JMC rule end-to-end testing
The idea is to provide java sources with code that violates
rules. We compile and execute the input sources while capturing
a JFR, and execute rules against the results to verify they work
as expected.

I like having the input sources available in the test rather than
only a JFR recording, because it's not clear whether most JDKs
will still create the given recording, or with what configuration.
However, the trade-off is that we're only testing against the
JVM used to run tests, which represents a subset of all supported
versions.
This type of test must also compile and execute small Java programs,
which is more expensive than executing against an existing recording,
and may be more likely to flake if we're not careful.

Anyhow, I thought I would share the idea before moving forward,
as always, feedback is appreciated :-)