Fix incorrect option name for `HeapDumpOnOutOfMemoryError`
I noticed that `HeapDumpOnOutOfMemoryError` was showing up as an
undocumented rule, and that seemed a bit surprising.

On a closer look, it's even documented on the exact link
https://docs.oracle.com/javase/8/docs/technotes/tools/windows/java.html
used as the source for the `JAVA_8_DOCUMENTED_XX`.

I googled and I don't think this option ever had another name, so
it looks like a copy-paste accident or something like that?

I've submitted the needed docs for the OCA so I'm opening the PR for
now and hopefully that will all be in good shape soon.